+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 ### Features
-- Added `-l/--local` flag to `armada ssh` which limits matched services to local machine.
 - Added service selection prompt to `armada ssh`. Whenever multiple matching services are found, user can easily select which instance they want to shh into.
+- Added `-n/--no-prompt` flag to `armada ssh` which disables prompting mechanism and results in error if multiple matching services are found. 
+- Added `-l/--local` flag to `armada ssh` which limits matched services to local machine.
 
 ### Bug fixes
 - Properly deregister crashed sub-services.